Add files via upload
REST API Sample to create a AWS Connection and update the Role ARN, SQS URL and Polling interval in a single call. 
Use this if you are automating the creation of AWS Connection in CWP with your AWS account from within your AWS infrastructure building scripts where an AWS account is created on the fly.